% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{auto_read}
\alias{auto_read}
\title{auto_read}
\usage{
auto_read(file, sep = ",", ...)
}
\arguments{
\item{file}{path to the file you want to import}

\item{sep}{which separator}

\item{...}{additionnal argument that can be passed on read function (see utils::read.csv() and readxl::read_excel())}
}
\description{
get the file extension and the separator values and call the adequate
function to read file.

ex : if a file was created with excel, the extension could either be
.xls or .xslx but would open with specific function.
The same applies to read.csv and read.csv2
}
\author{
Romain Garreau
}
