% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Subject_based_derivations.R
\name{basechar}
\alias{basechar}
\title{Baseline Covariates}
\usage{
basechar(df, group_col, date_col, ref_date_col, required_col, condition = NULL)
}
\arguments{
\item{condition}{}
}
\description{
Baseline Covariates
}
