% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{case_colname}
\alias{case_colname}
\title{case_colname}
\usage{
case_colname(data, case_type = c("lower", "upper"))
}
\arguments{
\item{data}{dataset provided}

\item{case_type}{Only lower or upper cases are accepted.}
}
\description{
wrapper function that help modify column name to upper or lower case in order to be case unsensitive
}
\author{
Romain Garreau
}
