% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Subject_based_derivations.R
\name{cg_creatinine_clearance}
\alias{cg_creatinine_clearance}
\title{creatinine in mg/dL in the formula}
\usage{
cg_creatinine_clearance(SEX, AGE, WEIGHT, WEIGHTU, CREAT, CREATU)
}
\arguments{
\item{CREATU}{}
}
\description{
creatinine in mg/dL in the formula
}
