% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_time_based_derivations.R
\name{compute_first_dose}
\alias{compute_first_dose}
\title{Title}
\usage{
compute_first_dose(df, subject_var = "USUBJID", date_var = "EXSTDTCN")
}
\arguments{
\item{date_var}{}
}
\description{
Title
}
