% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_time_based_derivations.R
\name{compute_lead}
\alias{compute_lead}
\title{Compute Lead In}
\usage{
compute_lead(
  df,
  subject_var = "USUBJID",
  date_var = "EXSTDTCN",
  lead_n = 1,
  new_var = "EXSTDTCN_LEAD"
)
}
\arguments{
\item{new_var}{}
}
\description{
Compute Lead In
}
