% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_time_based_derivations.R
\name{compute_TAD_from_PCTPT}
\alias{compute_TAD_from_PCTPT}
\title{Derive TAD}
\usage{
compute_TAD_from_PCTPT(
  data,
  pctpt_var = "PCTPT",
  pc_dt_var = "PCDTCN",
  ex_dt_var = "EXSTDTCN",
  ref_dt_var = "MEXSTDTCN",
  atafd_var = "ATAFD"
)
}
\arguments{
\item{atafd_var}{}
}
\description{
Derive TAD
}
