% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_time_based_derivations.R
\name{convert_datetime}
\alias{convert_datetime}
\title{Convert to numeric date time}
\usage{
convert_datetime(data, column)
}
\arguments{
\item{column}{}
}
\description{
Convert to numeric date time
}
