% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utl_create_metadata.R
\name{create_data_definition}
\alias{create_data_definition}
\title{Create metadata definition}
\usage{
create_data_definition(df, include_levels = TRUE, exclude_vars = NULL)
}
\arguments{
\item{exclude_vars}{}
}
\description{
Create metadata definition
}
