% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mb2_parse.R
\name{create_pk_data}
\alias{create_pk_data}
\title{create_pk_data}
\usage{
create_pk_data(admin_data, tdm_data, weight_data)
}
\arguments{
\item{admin_data}{A tibble with the administration data}

\item{tdm_data}{A tibble with the TDM data}

\item{weight_data}{A tibble with the weight data}
}
\value{
A tibble with the PK data
}
\description{
create_pk_data
}
\author{
Romain Garreau
}
