% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_time_based_derivations.R
\name{datenum}
\alias{datenum}
\title{Create numeric date and time variable}
\usage{
datenum(.data, date, type = "")
}
\arguments{
\item{.data}{dataset name}

\item{date}{date variable name}

\item{type}{date variable type}
}
\value{
convert character date time variable to numeric
}
\description{
Create numeric date and time variable
}
