% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_time_based_derivations.R
\name{derive_NTIME}
\alias{derive_NTIME}
\title{Derive NTIME}
\usage{
derive_NTIME(
  data,
  tad_var = "TAD",
  pc_dt_var = "PCDTCN",
  ex_dt_var = "EXSTDTCN",
  mex_dt_var = "MEXSTDTCN",
  fdose_var = "FDOSEDTN"
)
}
\arguments{
\item{fdose_var}{}
}
\description{
Derive NTIME
}
