% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_time_based_derivations.R
\name{join_and_compute_time}
\alias{join_and_compute_time}
\title{Actual Time Derivation}
\usage{
join_and_compute_time(
  df,
  join_df,
  by = "USUBJID",
  date_var = "EXSTDTCN",
  ref_var = "FDOSEDTN",
  extra_mutate = NULL
)
}
\arguments{
\item{extra_mutate}{}
}
\description{
Actual Time Derivation
}
