% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_pcex.R
\name{make_pcex}
\alias{make_pcex}
\title{Make PC EX}
\usage{
make_pcex(
  pc,
  ex,
  cols_to_fill = c("DOSFRM", "DOSFRQ", "ROUTE", "EXDOSE", "EXDOSU")
)
}
\arguments{
\item{cols_to_fill}{}
}
\description{
Make PC EX
}
