% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utl_create_numeric_codes.R
\name{map_columns_hybrid}
\alias{map_columns_hybrid}
\title{Create Numeric Codes For Covariates}
\usage{
map_columns_hybrid(
  df,
  user_orders = list(),
  auto_cols = NULL,
  start = 1,
  suffix = "N",
  overwrite = FALSE,
  drop_original = FALSE,
  default_code = 99L,
  missing_code = 0L,
  sort_auto = TRUE
)
}
\arguments{
\item{sort_auto}{}
}
\description{
Create Numeric Codes For Covariates
}
