% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mb2_parse.R
\name{mb2_parse}
\alias{mb2_parse}
\title{mb2_parse}
\usage{
mb2_parse(mb2_file)
}
\arguments{
\item{mb2_file}{a mb2_file}
}
\value{
A tibble with the formatted PK data
}
\description{
mb2_parse
}
\author{
Romain Garreau
}
