% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_mrgsolve.R
\name{mrg_check_colname}
\alias{mrg_check_colname}
\title{mrg_check_colname}
\usage{
mrg_check_colname(data, idat = FALSE)
}
\arguments{
\item{data}{data}

\item{idat}{Logical to specify if the dataset contains indivudal data or observation. Is False by default.}
}
\description{
check that colname are all adequate to use with mrg solve
}
\author{
Romain Garreau
}
