% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{not_in}
\alias{not_in}
\title{not_in}
\usage{
not_in(a, b)
}
\arguments{
\item{a}{value to check can be list or string}

\item{b}{value in which a is searched for, can be a single string or list of strings}
}
\description{
not_in = not in operator
}
\author{
Romain Garreau
}
