% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pksim_translate.R
\name{pm2psim}
\alias{pm2psim}
\title{pm2psim}
\usage{
pm2psim(
  x,
  sex = 0,
  time = "h",
  unit = "mg/L",
  tissue = "Plasma",
  compartment = "Peripheral Venous Blood",
  molecule = NULL,
  s_id = NULL,
  file = "Translate2pksim.csv"
)
}
\arguments{
\item{x}{= data set}

\item{sex}{= integer in 0:length(df), call number of the column corresponding to individual sex}

\item{time}{= time unit of the observation. by default time = "h"}

\item{unit}{= unit of measured concentration. by default "mg/L"}

\item{tissue}{correspond to the tissue from which the concentration was sampled, by default = 'Plasma'}

\item{compartment}{= compartment in which the sample have been taken. default value = "Peripheral Venous Blood"}

\item{molecule}{= molecule study, by default = NULL if none given}

\item{s_id}{= study id, by default = NULL}

\item{file}{= name given to the file. default = "translate.csv"}
}
\description{
This function allow to create a file ready to be used by PKsim
from a Pmetrics file.
}
