% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_pmetrics.R
\name{pm_autocomplete}
\alias{pm_autocomplete}
\title{pm_autocomplete}
\usage{
pm_autocomplete(data, missing_var, cov = NULL)
}
\arguments{
\item{data}{data from Pmetrics}

\item{missing_var}{missing variable. See ?PKtool::pm_check_colname}

\item{cov}{correspond to the covariate present in the dataset}
}
\description{
this function will add the necessary covariate in order to translate an incomplete Pmetrics file (Version 2 only)
}
\author{
Romain Garreau
}
