% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_pmetrics.R
\name{pm_check_colname}
\alias{pm_check_colname}
\title{pm_check_colname}
\usage{
pm_check_colname(data)
}
\arguments{
\item{data}{is the data in which the colum names must be checked}
}
\description{
function that check if the minimum colnames required to used Pmetrics are here.
}
