% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_domain.R
\name{process_domain}
\alias{process_domain}
\title{Process Domain}
\usage{
process_domain(
  file_name,
  projpath,
  datetime_var = NULL,
  convert_datetime_fun = convert_datetime,
  filter_expr = NULL,
  rename_list = NULL,
  drop_vars = NULL,
  mutate_expr = NULL
)
}
\arguments{
\item{mutate_expr}{}
}
\description{
Process Domain
}
