% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_finaldata.R
\name{process_final_data}
\alias{process_final_data}
\title{Derive Modeling Related Variables}
\usage{
process_final_data(
  df,
  study = 111,
  user_mutate = NULL,
  select_cols = keepvars,
  filter_expr = NULL
)
}
\arguments{
\item{filter_expr}{}
}
\description{
Derive Modeling Related Variables
}
