% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mb2_read.R
\name{read_file.mb2}
\alias{read_file.mb2}
\title{read_file.mb2}
\usage{
read_file.mb2(file_path)
}
\arguments{
\item{file_path}{correspond to the file location}
}
\description{
read_file.mb2 purpose is to get .mb2 file (PastrX file for BestDose (R)) and
extract all the useful information contained in the file
}
