% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mb2_read.R
\name{read_file}
\alias{read_file}
\title{read_file}
\usage{
read_file(file_path)
}
\arguments{
\item{file_path}{correspond to the file location}
}
\description{
read_file purpose is to identify the file extension and call the right subfunctions
to open and extract data from file
}
