% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_mrgsolve.R
\name{read_mrg}
\alias{read_mrg}
\title{read_mrg}
\usage{
read_mrg(data, ...)
}
\arguments{
\item{data}{mrgsolve format data type expected. Can be a path toward a file (csv, xls or xlsx) or a dataframe.}

\item{...}{additionnal argument possible see ?PKtool::auto_read}
}
\description{
read mrgsolve file types
}
\author{
Romain Garreau
}
