% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mb2_parse.R
\name{rearrange_admin_data}
\alias{rearrange_admin_data}
\title{rearrange_admin_data}
\usage{
rearrange_admin_data(admin_data)
}
\arguments{
\item{admin_data}{A tibble with the administration data}
}
\value{
A tibble with the rearranged administration data
}
\description{
rearrange_admin_data
}
\author{
Your Name
}
