% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mb2_parse.R
\name{rearrange_tdm_data}
\alias{rearrange_tdm_data}
\title{rearrange_tdm_data}
\usage{
rearrange_tdm_data(tdm_data)
}
\arguments{
\item{tdm_data}{A tibble with the TDM data}
}
\value{
A tibble with the rearranged TDM data
}
\description{
rearrange_tdm_data
}
\author{
Your Name
}
