% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mb2_parse.R
\name{rearrange_weight_data}
\alias{rearrange_weight_data}
\title{rearrange_weight_data}
\usage{
rearrange_weight_data(weight_data)
}
\arguments{
\item{weight_data}{A tibble with the weight data
@return A tibble with the rearranged weight data}
}
\description{
rearrange_weight_data
}
\author{
Romain Garreau
}
