% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utl_remap_codes.R
\name{remap_vec}
\alias{remap_vec}
\title{Recode Old Values To New Values}
\usage{
remap_vec(x, mapping, default = NA, missing_code = NA, warn_unmapped = TRUE)
}
\arguments{
\item{warn_unmapped}{}
}
\description{
Recode Old Values To New Values
}
