% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_renal_formula.R
\name{renal_function}
\alias{renal_function}
\title{renal_function}
\usage{
renal_function(
  sex,
  age,
  weight,
  height,
  creat,
  ethnicity = "Caucasian",
  formula,
  creat_unit = "uM/L",
  urine_creat = 0,
  urine_output = 0
)
}
\arguments{
\item{sex}{patient sex can be either 'Male' or 'Female'}

\item{age}{age of the patient}

\item{weight}{weight of the patient (can either be in kg or lbs)}

\item{height}{patient height.}

\item{creat}{serum creatinine can be given in mg/dL or uM/L}

\item{ethnicity}{can be African or other, is set to caucasian by default}

\item{formula}{the name of the formula needed to calculate the renal function. Can take any value in c('CG', 'MDRD', 'CKD_2009', 'CKD_2021', 'UVP', 'none').
These stand for Cockcroft and Gault, MDRD 4 variables, CKD-EPI 2009 and 2021 formula and calculated uv/p, respectively}

\item{creat_unit}{give the unit in which the creatinine was entered. Must be "uM/L" or "mg/dL". The unit is set to "uM/L" by default.}

\item{urine_creat}{the urine creatinine level. The unit is given by the parameter creat_unit}

\item{urine_output}{correspond to the daily urine output in mL. Note that this is the only unit that is supported}
}
\description{
renal_function
}
