% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utl_write_out.R
\name{save_study_data}
\alias{save_study_data}
\title{write out study data}
\usage{
save_study_data(study, substudy = NULL, final, projpath)
}
\arguments{
\item{projpath}{}
}
\description{
write out study data
}
