% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_file.R
\name{write_file}
\alias{write_file}
\title{write_file}
\usage{
write_file(object = NULL, filename = NULL, ...)
}
\arguments{
\item{object}{is the object supplied to write the file. Must be a dataframe or tibble}

\item{filename}{is the name the file should have}

\item{...}{dictate the type of file created (only file supported by pharmacometrics software are possible)}
}
\description{
this function create a file corresponding to the software for which the file was translated.
Only Pmetrics, Lixoft suite, PKsim and NONMEM type are supported. Software and extebsion type
are automatically retrvied from the object which must have been translated with the PKtool function.
Extention type is retrieved directly from the path given as the user must specify the path to save
the file by giving it name and extensiton ex : filename = "C:/newdata.csv"
}
\author{
Romain Garreau
}
