% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmetrics_translate.R
\name{pm2monolix}
\alias{pm2monolix}
\title{pm2monolix}
\usage{
pm2monolix(
  input_file = NULL,
  iv = TRUE,
  pm_vers = 1,
  output_file = "pm2monolix.csv",
  ...
)
}
\arguments{
\item{input_file}{datafile}

\item{iv}{is drug given by iv route ? (TRUE or FALSE)}

\item{pm_vers}{a numeric arugment that can be either 1 or 2. It refers to "old" or "new" Pmetrics version. Set to 1 by default (all Pmetrics version before 1.97)}

\item{output_file}{name given to the file. default = "translate.csv"}

\item{...}{additionnal argument that can be passed on read function (see utils::read.csv() and readxl::read_excel())}
}
\description{
'pm2monolix' will transform a pmetrics file into a file readable by Monolix(R)
}
\author{
Romain Garreau
}
