% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p2m.R
\name{pm2monolix}
\alias{pm2monolix}
\title{pm2monolix}
\usage{
pm2monolix(
  input_file = NULL,
  iv = TRUE,
  pm_vers = "old",
  output_file = "pm2monolix.csv"
)
}
\arguments{
\item{input_file}{datafile}

\item{iv}{is drug given by iv route ? (TRUE or FALSE)}

\item{pm_vers}{a string argument that can take either "old" or "new". Refer to the datafile version of
Pmetrics. Old represent the version before 1.97.0}

\item{output_file}{name given to the file. default = "translate.csv"}
}
\description{
'pm2monolix' is a function intended to transform a pm file into a file readable by Monolix(R)
}
