% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_pmetrics.R
\name{pm_autocomplete}
\alias{pm_autocomplete}
\title{pm_autocomplete}
\usage{
pm_autocomplete(data)
}
\arguments{
\item{data}{correspond to the data imported to run Pmetrics. Must be a dataframe}
}
\description{
this function will add the necessary covariate in order to translate an incomplete Pmetrics file (Version 2 only)
}
\author{
Romain Garreau
}
