% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_pmetrics.R
\name{read_pmetrics}
\alias{read_pmetrics}
\title{read_pmetrics}
\usage{
read_pmetrics(dataset, pm_vers = NULL)
}
\arguments{
\item{dataset}{is the name of the .csv file to read}

\item{pm_vers}{correspond to the version of Pmetrics.
Can take either 1 for all version before the 2.0.0 or 2 for all version above the 2.0.0}
}
\description{
this function purpose is to read and extract data from a pmetrics file.
It is the first step before any translation of a Pmetrics type file into
another software type.
}
