% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_file.R
\name{write_file}
\alias{write_file}
\title{write_file}
\usage{
write_file(
  object,
  filename,
  software = c("Pmetrics", "Lixoft", "PKsim", "NONMEM"),
  ext = c("csv", "xlsx", "xls", "tsv")
)
}
\arguments{
\item{object}{is the object supplied to write the file. Must be a dataframe or tibble}

\item{filename}{is the name the file should have}

\item{software}{correspond to the software for which the file is adapted. Only Pmetrics, Lixoft suite, PKsim and NONMEM are supported}

\item{ext}{dictate the type of file created
(only file supported by pharmacometrics software are possible)}
}
\description{
write_file
}
